Deactivator.php edited online with Bitbucket

+++ src/Deactivator.php
@@ -182,7 +182,7 @@ class Deactivator {
 		    'email'     	=> $user->user_email,
 		    'plugin'     	=> sanitize_text_field( $_POST['plugin'] ),
 		    'site_url'     	=> site_url( '/' ),
-		    'delay'     	=> time() - get_option( 'codesigner_install_time' ),
+		    'delay'     	=> time() - (int) get_option( 'codesigner_install_time' ),
 		    'reason'     	=> serialize( $_POST['reason'] ),
 		    'explanation'	=> sanitize_textarea_field( $_POST['explanation'] ),
 		], wp_unslash( $this->server ) );
